@@ -1,8 +1,8 @@
 module github.com/fsnotify/fsnotify
 
-go 1.16
+go 1.17
 
-require golang.org/x/sys v0.0.0-20220908164124-27713097b956
+require golang.org/x/sys v0.13.0
 
 retract (
 	v1.5.3 // Published an incorrect branch accidentally https://github.com/fsnotify/fsnotify/issues/445
