@@ -1,5 +1,4 @@
 //go:build solaris
-// +build solaris
 
 package fsnotify
 
@@ -24,24 +23,30 @@ func TestRemoveState(t *testing.T) {
 
 	check := func(wantDirs, wantFiles int) {
 		t.Helper()
-		if len(w.watches) != wantFiles {
+		if len(w.b.(*fen).watches) != wantFiles {
 			var d []string
-			for k, v := range w.watches {
+			for k, v := range w.b.(*fen).watches {
 				d = append(d, fmt.Sprintf("%#v = %#v", k, v))
 			}
 			t.Errorf("unexpected number of entries in w.watches (have %d, want %d):\n%v",
-				len(w.watches), wantFiles, strings.Join(d, "\n"))
+				len(w.b.(*fen).watches), wantFiles, strings.Join(d, "\n"))
 		}
-		if len(w.dirs) != wantDirs {
+		if len(w.b.(*fen).dirs) != wantDirs {
 			var d []string
-			for k, v := range w.dirs {
+			for k, v := range w.b.(*fen).dirs {
 				d = append(d, fmt.Sprintf("%#v = %#v", k, v))
 			}
 			t.Errorf("unexpected number of entries in w.dirs (have %d, want %d):\n%v",
-				len(w.dirs), wantDirs, strings.Join(d, "\n"))
+				len(w.b.(*fen).dirs), wantDirs, strings.Join(d, "\n"))
 		}
 	}
 
+	check(1, 1)
+
+	// Shouldn't change internal state.
+	if err := w.Add("/path-doesnt-exist"); err == nil {
+		t.Fatal(err)
+	}
 	check(1, 1)
 
 	if err := w.Remove(file); err != nil {
