@@ -7,45 +7,51 @@ import (
 	"path/filepath"
 	"runtime"
 	"sort"
+	"strconv"
 	"strings"
 	"sync"
 	"testing"
 	"time"
 
 	"github.com/fsnotify/fsnotify/internal"
+	"github.com/fsnotify/fsnotify/internal/ztest"
 )
-
-type testCase struct {
-	name string
-	ops  func(*testing.T, *Watcher, string)
-	want string
-}
-
-func (tt testCase) run(t *testing.T) {
-	t.Helper()
-	t.Run(tt.name, func(t *testing.T) {
-		t.Helper()
-		t.Parallel()
-		tmp := t.TempDir()
-
-		w := newCollector(t)
-		w.collect(t)
-
-		tt.ops(t, w.w, tmp)
-
-		cmpEvents(t, tmp, w.stop(t), newEvents(t, tt.want))
-	})
-}
 
 // We wait a little bit after most commands; gives the system some time to sync
 // things and makes things more consistent across platforms.
 func eventSeparator() { time.Sleep(50 * time.Millisecond) }
 func waitForEvents()  { time.Sleep(500 * time.Millisecond) }
 
+// To test the buffered watcher we run the tests twice in the CI: once as "go
+// test" and once with FSNOTIFY_BUFFER set. This is a bit hacky, but saves
+// having to refactor a lot of this code. Besides, running the tests in the CI
+// more than once isn't a bad thing, since it helps catch flaky tests (should
+// probably run it even more).
+var testBuffered = func() uint {
+	s, ok := os.LookupEnv("FSNOTIFY_BUFFER")
+	if ok {
+		i, err := strconv.ParseUint(s, 0, 0)
+		if err != nil {
+			panic(fmt.Sprintf("FSNOTIFY_BUFFER: %s", err))
+		}
+		return uint(i)
+	}
+	return 0
+}()
+
 // newWatcher initializes an fsnotify Watcher instance.
 func newWatcher(t *testing.T, add ...string) *Watcher {
 	t.Helper()
-	w, err := NewWatcher()
+
+	var (
+		w   *Watcher
+		err error
+	)
+	if testBuffered > 0 {
+		w, err = NewBufferedWatcher(testBuffered)
+	} else {
+		w, err = NewWatcher()
+	}
 	if err != nil {
 		t.Fatalf("newWatcher: %s", err)
 	}
@@ -67,6 +73,18 @@ func addWatch(t *testing.T, w *Watcher, path ...string) {
 	err := w.Add(join(path...))
 	if err != nil {
 		t.Fatalf("addWatch(%q): %s", join(path...), err)
+	}
+}
+
+// rmWatch removes a watch.
+func rmWatch(t *testing.T, watcher *Watcher, path ...string) {
+	t.Helper()
+	if len(path) < 1 {
+		t.Fatalf("rmWatch: path must have at least one element: %s", path)
+	}
+	err := watcher.Remove(join(path...))
+	if err != nil {
+		t.Fatalf("rmWatch(%q): %s", join(path...), err)
 	}
 }
 
@@ -105,13 +123,17 @@ func createFiles(t *testing.T, dir, prefix string, n int, d time.Duration) int {
 			t.Logf("createFiles: stopped at %s files because it took longer than %s", fmtNum(created), d)
 			return created
 		default:
-			fp, err := os.Create(join(dir, prefix+fmtNum(i)))
+			path := join(dir, prefix+fmtNum(i))
+			fp, err := os.Create(path)
 			if err != nil {
 				t.Errorf("create failed for %s: %s", fmtNum(i), err)
 				continue
 			}
 			if err := fp.Close(); err != nil {
 				t.Errorf("close failed for %s: %s", fmtNum(i), err)
+			}
+			if err := os.Remove(path); err != nil {
+				t.Errorf("remove failed for %s: %s", fmtNum(i), err)
 			}
 			if i%10_000 == 0 {
 				t.Logf("createFiles: %s", fmtNum(i))
@@ -138,19 +160,19 @@ func mkdir(t *testing.T, path ...string) {
 }
 
 // mkdir -p
-// func mkdirAll(t *testing.T, path ...string) {
-// 	t.Helper()
-// 	if len(path) < 1 {
-// 		t.Fatalf("mkdirAll: path must have at least one element: %s", path)
-// 	}
-// 	err := os.MkdirAll(join(path...), 0o0755)
-// 	if err != nil {
-// 		t.Fatalf("mkdirAll(%q): %s", join(path...), err)
-// 	}
-// 	if shouldWait(path...) {
-// 		eventSeparator()
-// 	}
-// }
+func mkdirAll(t *testing.T, path ...string) {
+	t.Helper()
+	if len(path) < 1 {
+		t.Fatalf("mkdirAll: path must have at least one element: %s", path)
+	}
+	err := os.MkdirAll(join(path...), 0o0755)
+	if err != nil {
+		t.Fatalf("mkdirAll(%q): %s", join(path...), err)
+	}
+	if shouldWait(path...) {
+		eventSeparator()
+	}
+}
 
 // ln -s
 func symlink(t *testing.T, target string, link ...string) {
@@ -197,15 +219,29 @@ func mknod(t *testing.T, dev int, path ...string) {
 	}
 }
 
-// cat
-func cat(t *testing.T, data string, path ...string) {
+// echo > and echo >>
+func echoAppend(t *testing.T, data string, path ...string) { t.Helper(); echo(t, false, data, path...) }
+func echoTrunc(t *testing.T, data string, path ...string)  { t.Helper(); echo(t, true, data, path...) }
+func echo(t *testing.T, trunc bool, data string, path ...string) {
+	n := "echoAppend"
+	if trunc {
+		n = "echoTrunc"
+	}
 	t.Helper()
 	if len(path) < 1 {
-		t.Fatalf("cat: path must have at least one element: %s", path)
+		t.Fatalf("%s: path must have at least one element: %s", n, path)
 	}
 
 	err := func() error {
-		fp, err := os.OpenFile(join(path...), os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
+		var (
+			fp  *os.File
+			err error
+		)
+		if trunc {
+			fp, err = os.Create(join(path...))
+		} else {
+			fp, err = os.OpenFile(join(path...), os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
+		}
 		if err != nil {
 			return err
 		}
@@ -227,7 +263,7 @@ func cat(t *testing.T, data string, path ...string) {
 		return fp.Close()
 	}()
 	if err != nil {
-		t.Fatalf("cat(%q): %s", join(path...), err)
+		t.Fatalf("%s(%q): %s", n, join(path...), err)
 	}
 }
 
@@ -290,6 +326,21 @@ func rmAll(t *testing.T, path ...string) {
 	err := os.RemoveAll(join(path...))
 	if err != nil {
 		t.Fatalf("rmAll(%q): %s", join(path...), err)
+	}
+	if shouldWait(path...) {
+		eventSeparator()
+	}
+}
+
+// cat
+func cat(t *testing.T, path ...string) {
+	t.Helper()
+	if len(path) < 1 {
+		t.Fatalf("cat: path must have at least one element: %s", path)
+	}
+	_, err := os.ReadFile(join(path...))
+	if err != nil {
+		t.Fatalf("cat(%q): %s", join(path...), err)
 	}
 	if shouldWait(path...) {
 		eventSeparator()
@@ -405,7 +456,11 @@ func (e Events) String() string {
 		if i > 0 {
 			b.WriteString("\n")
 		}
-		fmt.Fprintf(b, "%-20s %q", ee.Op.String(), filepath.ToSlash(ee.Name))
+		if ee.renamedFrom != "" {
+			fmt.Fprintf(b, "%-8s %s ← %s", ee.Op.String(), filepath.ToSlash(ee.Name), filepath.ToSlash(ee.renamedFrom))
+		} else {
+			fmt.Fprintf(b, "%-8s %s", ee.Op.String(), filepath.ToSlash(ee.Name))
+		}
 	}
 	return b.String()
 }
@@ -416,6 +471,11 @@ func (e Events) TrimPrefix(prefix string) Events {
 			e[i].Name = "/"
 		} else {
 			e[i].Name = strings.TrimPrefix(e[i].Name, prefix)
+		}
+		if e[i].renamedFrom == prefix {
+			e[i].renamedFrom = "/"
+		} else {
+			e[i].renamedFrom = strings.TrimPrefix(e[i].renamedFrom, prefix)
 		}
 	}
 	return e
@@ -472,44 +532,55 @@ func newEvents(t *testing.T, s string) Events {
 		}
 
 		fields := strings.Fields(line)
-		if len(fields) < 2 {
-			if strings.ToUpper(fields[0]) == "EMPTY" {
+		if len(fields) != 2 && len(fields) != 4 {
+			if strings.ToLower(fields[0]) == "empty" || strings.ToLower(fields[0]) == "no-events" {
 				for _, g := range groups {
 					events[g] = Events{}
 				}
 				continue
 			}
-
-			t.Fatalf("newEvents: line %d has less than 2 fields: %s", no, line)
+			t.Fatalf("newEvents: line %d: needs 2 or 4 fields: %s", no+1, line)
 		}
 
-		path := strings.Trim(fields[len(fields)-1], `"`)
-
 		var op Op
-		for _, e := range fields[:len(fields)-1] {
-			if e == "|" {
-				continue
+		for _, ee := range strings.Split(fields[0], "|") {
+			switch strings.ToUpper(ee) {
+			case "CREATE":
+				op |= Create
+			case "WRITE":
+				op |= Write
+			case "REMOVE":
+				op |= Remove
+			case "RENAME":
+				op |= Rename
+			case "CHMOD":
+				op |= Chmod
+			case "OPEN":
+				op |= xUnportableOpen
+			case "READ":
+				op |= xUnportableRead
+			case "CLOSE_WRITE":
+				op |= xUnportableCloseWrite
+			case "CLOSE_READ":
+				op |= xUnportableCloseRead
+			default:
+				t.Fatalf("newEvents: line %d has unknown event %q: %s", no+1, ee, line)
 			}
-			for _, ee := range strings.Split(e, "|") {
-				switch strings.ToUpper(ee) {
-				case "CREATE":
-					op |= Create
-				case "WRITE":
-					op |= Write
-				case "REMOVE":
-					op |= Remove
-				case "RENAME":
-					op |= Rename
-				case "CHMOD":
-					op |= Chmod
-				default:
-					t.Fatalf("newEvents: line %d has unknown event %q: %s", no, ee, line)
-				}
+		}
+
+		var from string
+		if len(fields) > 2 {
+			if fields[2] != "←" {
+				t.Fatalf("newEvents: line %d: invalid format: %s", no+1, line)
 			}
+			from = strings.Trim(fields[3], `"`)
+		}
+		if !supportsRename() {
+			from = ""
 		}
 
 		for _, g := range groups {
-			events[g] = append(events[g], Event{Name: path, Op: op})
+			events[g] = append(events[g], Event{Name: strings.Trim(fields[1], `"`), renamedFrom: from, Op: op})
 		}
 	}
 
@@ -545,8 +616,9 @@ func cmpEvents(t *testing.T, tmp string, have, want Events) {
 	})
 
 	if haveSort.String() != wantSort.String() {
-		//t.Error("\n" + ztest.Diff(indent(haveSort), indent(wantSort)))
-		t.Errorf("\nhave:\n%s\nwant:\n%s", indent(have), indent(want))
+		b := new(strings.Builder)
+		b.WriteString(strings.TrimSpace(ztest.Diff(indent(haveSort), indent(wantSort))))
+		t.Errorf("\nhave:\n%s\nwant:\n%s\ndiff:\n%s", indent(have), indent(want), indent(b))
 	}
 }
 
@@ -555,11 +627,6 @@ func indent(s fmt.Stringer) string {
 }
 
 var join = filepath.Join
-
-func isCI() bool {
-	_, ok := os.LookupEnv("CI")
-	return ok
-}
 
 func isKqueue() bool {
 	switch runtime.GOOS {
@@ -575,4 +642,382 @@ func isSolaris() bool {
 		return true
 	}
 	return false
+}
+
+func supportsRecurse(t *testing.T) {
+	switch runtime.GOOS {
+	case "windows", "linux":
+		// Run test.
+	default:
+		t.Skip("recursion not yet supported on " + runtime.GOOS)
+	}
+}
+
+func supportsFilter(t *testing.T) {
+	switch runtime.GOOS {
+	case "linux":
+		// Run test.
+	default:
+		t.Skip("withOps() not yet supported on " + runtime.GOOS)
+	}
+}
+
+func supportsRename() bool {
+	switch runtime.GOOS {
+	case "linux", "windows":
+		return true
+	default:
+		return false
+	}
+}
+
+func supportsNofollow(t *testing.T) {
+	switch runtime.GOOS {
+	case "linux":
+		// Run test.
+	default:
+		t.Skip("withNoFollow() not yet supported on " + runtime.GOOS)
+	}
+}
+
+func tmppath(tmp, s string) string {
+	if len(s) == 0 {
+		return ""
+	}
+	if !strings.HasPrefix(s, "./") {
+		return filepath.Join(tmp, s)
+	}
+	// Needed for creating relative links. Support that only with explicit "./"
+	// – otherwise too easy to forget leading "/" and create files outside of
+	// the tmp dir.
+	return s
+}
+
+type command struct {
+	line int
+	cmd  string
+	args []string
+}
+
+func parseScript(t *testing.T, in string) {
+	var (
+		lines = strings.Split(in, "\n")
+		cmds  = make([]command, 0, 8)
+		readW bool
+		want  string
+		tmp   = t.TempDir()
+	)
+	for i, line := range lines {
+		line = strings.TrimSpace(line)
+		if line == "" || line[0] == '#' {
+			continue
+		}
+		if i := strings.IndexByte(line, '#'); i > -1 {
+			line = strings.TrimSpace(line[:i])
+		}
+		if line == "Output:" {
+			readW = true
+			continue
+		}
+		if readW {
+			want += line + "\n"
+			continue
+		}
+
+		cmd := command{line: i + 1, args: make([]string, 0, 4)}
+		var (
+			q   bool
+			cur = make([]rune, 0, 16)
+			app = func() {
+				if len(cur) == 0 {
+					return
+				}
+				if cmd.cmd == "" {
+					cmd.cmd = string(cur)
+				} else {
+					cmd.args = append(cmd.args, string(cur))
+				}
+				cur = cur[:0]
+			}
+		)
+		for _, c := range line {
+			switch c {
+			case ' ', '\t':
+				if q {
+					cur = append(cur, c)
+				} else {
+					app()
+				}
+			case '"', '\'': // '
+				q = !q
+			default:
+				cur = append(cur, c)
+			}
+		}
+		app()
+		cmds = append(cmds, cmd)
+	}
+
+	var (
+		do      = make([]func(), 0, len(cmds))
+		w       = newCollector(t)
+		mustArg = func(c command, n int) {
+			if len(c.args) != n {
+				t.Fatalf("line %d: %q requires exactly %d argument (have %d: %q)",
+					c.line, c.cmd, n, len(c.args), c.args)
+			}
+		}
+	)
+loop:
+	for _, c := range cmds {
+		c := c
+		//fmt.Printf("line %d: %q  %q\n", c.line, c.cmd, c.args)
+		switch c.cmd {
+		case "skip", "require":
+			mustArg(c, 1)
+			switch c.args[0] {
+			case "op_all":
+				if runtime.GOOS != "linux" {
+					t.Skip("No op_all on this platform")
+				}
+			case "op_open":
+				if runtime.GOOS != "linux" {
+					t.Skip("No Open on this platform")
+				}
+			case "op_read":
+				if runtime.GOOS != "linux" {
+					t.Skip("No Read on this platform")
+				}
+			case "op_close_write":
+				if runtime.GOOS != "linux" {
+					t.Skip("No CloseWrite on this platform")
+				}
+			case "op_close_read":
+				if runtime.GOOS != "linux" {
+					t.Skip("No CloseRead on this platform")
+				}
+			case "always":
+				t.Skip()
+			case "symlink":
+				if !internal.HasPrivilegesForSymlink() {
+					t.Skipf("%s symlink: admin permissions required on Windows", c.cmd)
+				}
+			case "mkfifo":
+				if runtime.GOOS == "windows" {
+					t.Skip("No named pipes on Windows")
+				}
+			case "mknod":
+				if runtime.GOOS == "windows" {
+					t.Skip("No device nodes on Windows")
+				}
+				if isKqueue() {
+					// Don't want to use os/user to check uid, since that pulls
+					// in cgo by default and stuff that uses fsnotify won't be
+					// statically linked by default.
+					t.Skip("needs root on BSD")
+				}
+				if isSolaris() {
+					t.Skip(`"mknod fails with "not owner"`)
+				}
+			case "recurse":
+				supportsRecurse(t)
+			case "filter":
+				supportsFilter(t)
+			case "nofollow":
+				supportsNofollow(t)
+			case "windows":
+				if runtime.GOOS == "windows" {
+					t.Skip("Skipping on Windows")
+				}
+			case "netbsd":
+				if runtime.GOOS == "netbsd" {
+					t.Skip("Skipping on NetBSD")
+				}
+			case "openbsd":
+				if runtime.GOOS == "openbsd" {
+					t.Skip("Skipping on OpenBSD")
+				}
+			default:
+				t.Fatalf("line %d: unknown %s reason: %q", c.line, c.cmd, c.args[0])
+			}
+		//case "state":
+		//	mustArg(c, 0)
+		//	do = append(do, func() { eventSeparator(); fmt.Fprintln(os.Stderr); w.w.state(); fmt.Fprintln(os.Stderr) })
+		case "debug":
+			mustArg(c, 1)
+			switch c.args[0] {
+			case "1", "on", "true", "yes":
+				do = append(do, func() { debug = true })
+			case "0", "off", "false", "no":
+				do = append(do, func() { debug = false })
+			default:
+				t.Fatalf("line %d: unknown debug: %q", c.line, c.args[0])
+			}
+		case "stop":
+			mustArg(c, 0)
+			break loop
+		case "watch":
+			if len(c.args) < 1 {
+				t.Fatalf("line %d: %q requires at least %d arguments (have %d: %q)",
+					c.line, c.cmd, 1, len(c.args), c.args)
+			}
+			if len(c.args) == 1 {
+				do = append(do, func() { addWatch(t, w.w, tmppath(tmp, c.args[0])) })
+				continue
+			}
+
+			var follow addOpt
+			for i := range c.args {
+				if c.args[i] == "nofollow" || c.args[i] == "no-follow" {
+					c.args = append(c.args[:i], c.args[i+1:]...)
+					follow = withNoFollow()
+					break
+				}
+			}
+
+			var op Op
+			for _, o := range c.args[1:] {
+				switch strings.ToLower(o) {
+				default:
+					t.Fatalf("line %d: unknown: %q", c.line+1, o)
+				case "default":
+					op |= Create | Write | Remove | Rename | Chmod
+				case "create":
+					op |= Create
+				case "write":
+					op |= Write
+				case "remove":
+					op |= Remove
+				case "rename":
+					op |= Rename
+				case "chmod":
+					op |= Chmod
+				case "open":
+					op |= xUnportableOpen
+				case "read":
+					op |= xUnportableRead
+				case "close_write":
+					op |= xUnportableCloseWrite
+				case "close_read":
+					op |= xUnportableCloseRead
+				}
+			}
+			do = append(do, func() {
+				p := tmppath(tmp, c.args[0])
+				err := w.w.AddWith(p, withOps(op), follow)
+				if err != nil {
+					t.Fatalf("line %d: addWatch(%q): %s", c.line+1, p, err)
+				}
+			})
+		case "unwatch":
+			mustArg(c, 1)
+			do = append(do, func() { rmWatch(t, w.w, tmppath(tmp, c.args[0])) })
+		case "watchlist":
+			mustArg(c, 1)
+			n, err := strconv.ParseInt(c.args[0], 10, 0)
+			if err != nil {
+				t.Fatalf("line %d: %s", c.line, err)
+			}
+			do = append(do, func() {
+				wl := w.w.WatchList()
+				if l := int64(len(wl)); l != n {
+					t.Errorf("line %d: watchlist has %d entries, not %d\n%q", c.line, l, n, wl)
+				}
+			})
+		case "touch":
+			mustArg(c, 1)
+			do = append(do, func() { touch(t, tmppath(tmp, c.args[0])) })
+		case "mkdir":
+			recur := false
+			if len(c.args) == 2 && c.args[0] == "-p" {
+				recur, c.args = true, c.args[1:]
+			}
+			mustArg(c, 1)
+			if recur {
+				do = append(do, func() { mkdirAll(t, tmppath(tmp, c.args[0])) })
+			} else {
+				do = append(do, func() { mkdir(t, tmppath(tmp, c.args[0])) })
+			}
+		case "ln":
+			mustArg(c, 3)
+			if c.args[0] != "-s" {
+				t.Fatalf("line %d: only ln -s is supported", c.line)
+			}
+			do = append(do, func() { symlink(t, tmppath(tmp, c.args[1]), tmppath(tmp, c.args[2])) })
+		case "mkfifo":
+			mustArg(c, 1)
+			do = append(do, func() { mkfifo(t, tmppath(tmp, c.args[0])) })
+		case "mknod":
+			mustArg(c, 2)
+			n, err := strconv.ParseInt(c.args[0], 10, 0)
+			if err != nil {
+				t.Fatalf("line %d: %s", c.line, err)
+			}
+			do = append(do, func() { mknod(t, int(n), tmppath(tmp, c.args[1])) })
+		case "mv":
+			mustArg(c, 2)
+			do = append(do, func() { mv(t, tmppath(tmp, c.args[0]), tmppath(tmp, c.args[1])) })
+		case "rm":
+			recur := false
+			if len(c.args) == 2 && c.args[0] == "-r" {
+				recur, c.args = true, c.args[1:]
+			}
+			mustArg(c, 1)
+			if recur {
+				do = append(do, func() { rmAll(t, tmppath(tmp, c.args[0])) })
+			} else {
+				do = append(do, func() { rm(t, tmppath(tmp, c.args[0])) })
+			}
+		case "chmod":
+			mustArg(c, 2)
+			n, err := strconv.ParseUint(c.args[0], 8, 32)
+			if err != nil {
+				t.Fatalf("line %d: %s", c.line, err)
+			}
+			do = append(do, func() { chmod(t, fs.FileMode(n), tmppath(tmp, c.args[1])) })
+		case "cat":
+			mustArg(c, 1)
+			do = append(do, func() { cat(t, tmppath(tmp, c.args[0])) })
+		case "echo":
+			if len(c.args) < 2 || len(c.args) > 3 {
+				t.Fatalf("line %d: %q requires 2 or 3 arguments (have %d: %q)",
+					c.line, c.cmd, len(c.args), c.args)
+			}
+
+			var data, op, dst string
+			if len(c.args) == 2 { // echo foo >dst
+				data, op, dst = c.args[0], c.args[1][:1], c.args[1][1:]
+				if strings.HasPrefix(dst, ">") {
+					op, dst = op+dst[:1], dst[1:]
+				}
+			} else { // echo foo > dst
+				data, op, dst = c.args[0], c.args[1], c.args[2]
+			}
+
+			switch op {
+			case ">":
+				do = append(do, func() { echoTrunc(t, data, tmppath(tmp, dst)) })
+			case ">>":
+				do = append(do, func() { echoAppend(t, data, tmppath(tmp, dst)) })
+			default:
+				t.Fatalf("line %d: echo requires > (truncate) or >> (append): echo data >file", c.line)
+			}
+		case "sleep":
+			mustArg(c, 1)
+			n, err := strconv.ParseInt(strings.TrimRight(c.args[0], "ms"), 10, 0)
+			if err != nil {
+				t.Fatalf("line %d: %s", c.line, err)
+			}
+			do = append(do, func() { time.Sleep(time.Duration(n) * time.Millisecond) })
+		default:
+			t.Errorf("line %d: unknown command %q", c.line, c.cmd)
+		}
+	}
+
+	w.collect(t)
+	for _, d := range do {
+		d()
+	}
+	ev := w.stop(t)
+	cmpEvents(t, tmp, ev, newEvents(t, want))
 }
