@@ -27,11 +27,19 @@ func Debug(name string, mask int32) {
 		{"FILE_EXCEPTION", unix.FILE_EXCEPTION},
 	}
 
-	var l []string
+	var (
+		l       []string
+		unknown = mask
+	)
 	for _, n := range names {
 		if mask&n.m == n.m {
 			l = append(l, n.n)
+			unknown ^= n.m
 		}
 	}
-	fmt.Fprintf(os.Stderr, "%s  %-20s → %s\n", time.Now().Format("15:04:05.0000"), strings.Join(l, " | "), name)
+	if unknown > 0 {
+		l = append(l, fmt.Sprintf("0x%x", unknown))
+	}
+	fmt.Fprintf(os.Stderr, "FSNOTIFY_DEBUG: %s  %10d:%-30s → %q\n",
+		time.Now().Format("15:04:05.000000000"), mask, strings.Join(l, " | "), name)
 }
