@@ -63,7 +63,7 @@ int main(int argc, char* argv[]) {
 		char *path = malloc(strlen(dir) + strlen(ls->d_name) + 2);
 		sprintf(path, "%s/%s", dir, ls->d_name);
 
-		int fp = open(path, O_RDONLY);
+		int fp = open(path, O_RDONLY | O_PATH | O_NOFOLLOW);
 		if (fp == -1)
 			die("open: %s:", path);
 		fds[++n_fds] = fp;
